// lib/api/types.gen.ts
export type ProductionBonusEditModel = {
    id?: number;
    startDate?: string;
    stopDate?: string;
    aCrew?: number;
    bCrew?: number;
    cCrew?: number;
    dCrew?: number;
    average?: number;
};